[WebNN] Save NaN instead of 0 for MLNumber when initialized from NaN.
This CL saves NaN instead of 0 for MLNumber when initialized from
NaN. The behavior on different backends and hardwares are completely
inconsistent when NaN is used for min_value or max_value of clamp.
So we normalize it with minValue:NaN coerced to -Infinity and
maxValue:NaN be coerced to Infinity.

Bug: 335909582
Change-Id: I4c7467b573010f7804438dc1d2993e45594b0708