[gap-decorations] Fix underlying GapDataListInterpolationType Checkers
This CL fixes the Underlying checkers for both GapLengthList and
GapColorList. Other InterpolationTypes (such as LengthList and Color)
don't use these either for `MaybeConvertNeutral`, and they were leading
to some fuzzer crashes since `Equals()` are marked as `NOTREACHED` in
`BaseInterpolableColor` and `InterpolableLength`.

The way we fix these is by making the `IsValid` check to simply check
that the lengths match and that he values in the lists are compatible.

Bug: 357648037
Fixed: 432924205,431949988
Change-Id: I50d0157b5514c80b0cc1b2673faf0536782007f4